cascade metadata for document title
